wip: better caddy config reconciliation
