+++ caddy/caddy.go
@@ -1,14 +1,10 @@
 package caddy
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
-	"hash/fnv"
 	"io"
-	"net/http"
-	"net/url"
 	"os"
 	"sort"
 	"strconv"
@@ -26,7 +22,6 @@ import (
 	"github.com/krystal/guvnor/ready"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
-	"k8s.io/apimachinery/pkg/util/rand"
 )
 
 const (
@@ -61,48 +56,52 @@ type Manager struct {
 	ContainerLabels map[string]string
 }
 
-func hashConfig(cfg *caddy.Config) string {
-	hasher := fnv.New32()
-	fmt.Fprintf(hasher, "%#v", cfg)
-	return rand.SafeEncodeString(fmt.Sprint(hasher.Sum32()))
-}
+func (cm *Manager) calculateConfigChanges(config *caddy.Config) (bool, error) {
+	// TODO: Swap this out for some hashing or string comparison
+	// I originally tried this with a JSONified version of the config, but
+	// unfortunately this did not work as the keys changed position. This could
+	// be worked around but in the interest of time, I went with a bool.
+	hasChanged := false
 
-func (cm *Manager) reconcileCaddyConfig(ctx context.Context) error {
-	config, err := cm.getConfig(ctx)
-	if err != nil {
-		return err
+	if config.AppsRaw == nil {
+		config.AppsRaw = caddy.ModuleMap{}
+		hasChanged = true
 	}
-	initialHash := hashConfig(config)
 
 	httpConfig := &caddyhttp.App{}
 	currentHTTPConfigRaw, ok := config.AppsRaw["http"]
 	if ok {
 		if err := json.Unmarshal(currentHTTPConfigRaw, httpConfig); err != nil {
-			return err
+			return hasChanged, err
 		}
 	}
 
 	if httpConfig.HTTPPort != cm.Config.Ports.HTTP {
 		httpConfig.HTTPPort = cm.Config.Ports.HTTP
+		hasChanged = true
 	}
 
 	if httpConfig.HTTPSPort != cm.Config.Ports.HTTPS {
 		httpConfig.HTTPSPort = cm.Config.Ports.HTTPS
+		hasChanged = true
 	}
 
 	if httpConfig.Servers == nil {
 		httpConfig.Servers = map[string]*caddyhttp.Server{}
+		hasChanged = true
 	}
 
 	serverConfig := httpConfig.Servers[guvnorServerName]
 	if serverConfig == nil {
 		serverConfig = &caddyhttp.Server{}
 		httpConfig.Servers[guvnorServerName] = serverConfig
+		hasChanged = true
 	}
 
 	listenAddr := ":" + strconv.Itoa(cm.Config.Ports.HTTPS)
 	if len(serverConfig.Listen) != 1 || serverConfig.Listen[0] != listenAddr {
 		serverConfig.Listen = []string{listenAddr}
+		hasChanged = true
 	}
 
 	// TODO: Be a bit smarter and create/update the default route
@@ -115,7 +114,7 @@ func (cm *Manager) reconcileCaddyConfig(ctx context.Context) error {
 		}
 		defaultHandlerRaw, err := json.Marshal(defaultHandler)
 		if err != nil {
-			return err
+			return hasChanged, err
 		}
 
 		serverConfig.Routes = append(serverConfig.Routes,
@@ -125,6 +124,7 @@ func (cm *Manager) reconcileCaddyConfig(ctx context.Context) error {
 				},
 			},
 		)
+		hasChanged = true
 	}
 
 	// TODO: Clean up servers we don't recognise ??
@@ -132,19 +132,27 @@ func (cm *Manager) reconcileCaddyConfig(ctx context.Context) error {
 	// Persist the HTTP config to the main config
 	modifiedHTTPConfigRaw, err := json.Marshal(httpConfig)
 	if err != nil {
-		return err
+		return hasChanged, err
 	}
 	config.AppsRaw["http"] = modifiedHTTPConfigRaw
 
-	finalHash := hashConfig(config)
-	// Compare hash of confeeg ?
-	if initialHash != finalHash {
-		err = cm.doRequest(
-			ctx, http.MethodPost, &url.URL{Path: "config/"}, config, nil,
-		)
-		if err != nil {
-			return err
-		}
+	return hasChanged, err
+}
+
+func (cm *Manager) reconcileCaddyConfig(ctx context.Context) error {
+	config, err := cm.getConfig(ctx)
+	if err != nil {
+		return err
+	}
+
+	hasChanged, err := cm.calculateConfigChanges(config)
+	if err != nil {
+		return err
+	}
+
+	if hasChanged {
+		cm.Log.Info("reconciliation found changes, updating caddy config")
+		return cm.postConfig(ctx, config)
 	}
 
 	return nil
@@ -353,111 +361,4 @@ func (cm *Manager) ConfigureBackend(
 	sortRoutes(routes)
 
 	return cm.patchRoutes(ctx, routes)
-}
-
-// getRoutes returns an slice of routes configured on the caddy server
-func (cm *Manager) getRoutes(ctx context.Context) ([]route, error) {
-	currentRoutes := []route{}
-	routesConfigPath := fmt.Sprintf(
-		"config/apps/http/servers/%s/routes",
-		guvnorServerName,
-	)
-	err := cm.doRequest(ctx, http.MethodGet, &url.URL{Path: routesConfigPath}, nil, &currentRoutes)
-	if err != nil {
-		return nil, err
-	}
-
-	return currentRoutes, nil
-}
-
-// prependRoute adds a new route to the start of the route array in the server
-func (cm *Manager) patchRoutes(ctx context.Context, route []route) error {
-	prependRoutePath := fmt.Sprintf(
-		"config/apps/http/servers/%s/routes",
-		guvnorServerName,
-	)
-	return cm.doRequest(
-		ctx,
-		http.MethodPatch,
-		&url.URL{Path: prependRoutePath},
-		route,
-		nil,
-	)
-}
-
-func (cm *Manager) getConfig(ctx context.Context) (*caddy.Config, error) {
-	cfg := &caddy.Config{}
-	err := cm.doRequest(
-		ctx,
-		http.MethodGet,
-		&url.URL{Path: "config/"},
-		nil,
-		cfg,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	return cfg, nil
-}
-
-func (cm *Manager) doRequest(ctx context.Context, method string, path *url.URL, body interface{}, out interface{}) error {
-	var bodyToSend io.Reader
-	if body != nil {
-		if v, ok := body.(string); ok {
-			// Send string directly
-			bodyToSend = bytes.NewBufferString(v)
-		} else if v, ok := body.([]byte); ok {
-			bodyToSend = bytes.NewBuffer(v)
-		} else {
-			// If not a string, JSONify it and send it
-			data, err := json.Marshal(body)
-			if err != nil {
-				return fmt.Errorf("marshalling body: %w", err)
-			}
-			bodyToSend = bytes.NewBuffer(data)
-		}
-	}
-
-	// TODO: Pull this into the config for Manager
-	rootPath, err := url.Parse("http://localhost:2019")
-	if err != nil {
-		return err
-	}
-
-	fullPath := rootPath.ResolveReference(path).String()
-
-	req, err := http.NewRequestWithContext(ctx, method, fullPath, bodyToSend)
-	if err != nil {
-		return fmt.Errorf("creating request: %w", err)
-	}
-
-	req.Header.Add("Content-Type", "application/json")
-
-	cm.Log.Debug("making request to caddy",
-		zap.String("url", req.URL.String()),
-		zap.String("method", req.Method),
-	)
-	res, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return fmt.Errorf("executing request: %w", err)
-	}
-	defer res.Body.Close()
-
-	// TODO: Check status codes
-	data, err := io.ReadAll(res.Body)
-	if err != nil {
-		return fmt.Errorf("reading response body: %w", err)
-	}
-	cm.Log.Debug("response from caddy",
-		zap.String("body", string(data)),
-		zap.Int("status", res.StatusCode),
-	)
-	if out != nil {
-		if err := json.Unmarshal(data, out); err != nil {
-			return fmt.Errorf("unmarshalling response: %w", err)
-		}
-	}
-
-	return nil
 }
